requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
mysql-connector-python==8.1.0
Werkzeug==2.3.7
python-dotenv==1.0.0
Pillow==11.3.0 